Fixed warning

+++ Pyro/src/pyro/graphics/raw_model.cpp
@@ -1,15 +1,15 @@
 ﻿#include "pyro_pch.h"
 #include "raw_model.h"
 
-pyro::raw_model::raw_model(unsigned int vao_id, int vertex_count)
+pyro::raw_model::raw_model(unsigned int vao_id, size_t vertex_count)
 {
     create(vao_id, vertex_count);
 }
 
-void pyro::raw_model::create(unsigned int vao_id, int vertex_count)
+void pyro::raw_model::create(unsigned int vao_id, size_t vertex_count)
 {
     m_vao_id = vao_id;
-    m_vertex_count = vertex_count;
+    m_vertex_count = static_cast<int>(vertex_count);
 }
 
 unsigned pyro::raw_model::vao_id() const
